add authorization policies for MusicSet (aka epic security added :heavy_check_mark:)

+++ site/app/Http/Controllers/MusicController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Http\Requests\StoreMusicSet;
+use App\Http\Requests\MusicSetRequest;
 use App\MusicSet;
 use App\MusicFile;
 use Auth;
@@ -13,6 +13,7 @@ class MusicController extends Controller
     /**
      * Display a listing of the resource.
      *
+     * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
     public function index(Request $request)
@@ -29,17 +30,25 @@ class MusicController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->can('create', MusicSet::class)) {
+            return abort(403);
+        }
+
         return view('music.create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Requests\MusicSetRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(StoreMusicSet $request)
+    public function store(MusicSetRequest $request)
     {
+        if (!Auth::user()->can('store', MusicSet::class)) {
+            return abort(403);
+        }
+
         if ($request->file('music_file')->isValid()) {
             $musicFilePath = $request->file('music_file')->store('music_files');
         }
@@ -65,12 +74,18 @@ class MusicController extends Controller
     /**
      * Display the specified resource.
      *
+     * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function show(Request $request, $id)
     {
         $musicSet = MusicSet::findOrFail($id);
+
+        if (!Auth::user()->can('view', $musicSet)) {
+            return abort(403);
+        }
+
         $musicFiles = $musicSet->musicFiles;
 
         return view('music.show', compact('musicSet', 'musicFiles'));
@@ -79,6 +94,7 @@ class MusicController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
+     * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
@@ -86,18 +102,28 @@ class MusicController extends Controller
     {
         $musicSet = MusicSet::findOrFail($id);
 
+        if (!Auth::user()->can('edit', $musicSet)) {
+            abort(403);
+        }
+
         return view('music.edit', compact('musicSet'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Requests\MusicSetRequest  $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(StoreMusicSet $request, $id)
+    public function update(MusicSetRequest $request, $id)
     {
+        $musicSet = MusicSet::findOrFail($id);
+
+        if (!Auth::user()->can('update', $musicSet)) {
+            abort(403);
+        }
+
         if ($request->file('music_file')->isValid()) {
             $musicFilePath = $request->file('music_file')->store('music_files');
         }
@@ -105,11 +131,10 @@ class MusicController extends Controller
             return redirect()->back();
         }
 
-        $musicSet = MusicSet::find($id);
         $musicSet->name = $request->music_set_name;
         $musicSet->save();
 
-        $musicFile = MusicFile::where('music_set_id', $musicSet->id)->first();
+        $musicFile = MusicFile::where('music_set_id', $musicSet->id)->first(); // ->first() is temp
         $musicFile->path = $musicFilePath;
         $musicFile->save();
 
@@ -121,14 +146,19 @@ class MusicController extends Controller
     /**
      * Remove the specified resource from storage.
      *
+     * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function destroy(Request $request, $id)
     {
-        $musicSet = MusicSet::find($id);
-        $musicSet->delete();
+        $musicSet = MusicSet::findOrFail($id);
 
+        if (!Auth::user()->can('destroy', $musicSet)) {
+            abort(403);
+        }
+
+        $musicSet->delete();
         $request->session()->flash('success', 'Successfully deleted!');
 
         return redirect()->route('music.index');
